backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
pymysql==1.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
python-multipart==0.0.6
google-search-results==2.4.2
httpx==0.25.2
openai==1.3.7
pillow==10.1.0
python-dotenv==1.0.0
cors==1.0.1
email-validator==2.1.0
redis==5.0.1
celery==5.3.4
boto3==1.34.0
requests==2.31.0
